make autotask source code network agnostic
