--- essential-nft-meta-tx/src/index.d.ts
@@ -1,10 +1,0 @@
-interface ForwardRequest {
-  to: string;
-  from: string;
-  authorizer: string;
-  nftContract: string;
-  nonce: string;
-  nftNonce: string;
-  tokenId: string;
-  data: string;
-}
